+++ src/index.iife.ts
@@ -1,0 +1,1 @@
+export { tokenizer as default } from './index.js'