+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changes to CSS Tokenizer
 
+### 2.0.0 (May 11, 2021)
+
+- Supports function token (previously an identifier token followed by a parenthesis delimiter token).
+- Changed from default export to named export (`tokenizer`).
+
 ### 1.0.0 (September 26, 2020)
 
 Initial version